Create spec system of deleting expression when next expression does not exist and none of expresisons exist
